Started HW02; added essay on FP
I started going through the second week of homework for cis194, and had
a realization. I had seen the light that is functional programming and
needed to write an essay about my experience. The essay is to remind
myself of what can be accomplished with extreme elegance in a functional
style (especially in Haskell).